+++ src/modules/bin/about_module.rs
@@ -6,11 +6,11 @@ use common_macros::hash_map;
 
 pub fn get() -> Expression {
     (hash_map! {
-        String::from("author") => Expression::String("Adam McDaniel, santo".to_string()),
+        String::from("author") => Expression::String("Santo, Adam McDaniel".to_string()),
         String::from("git") => Expression::String("https://codeberg.com/santo/lumesh".to_string()),
         String::from("homepage") => Expression::String("https://lumesh.codeberg.page".to_string()),
         String::from("version") => Expression::String(VERSION.to_string()),
-        String::from("path") => {
+        String::from("bin") => {
             if let Ok(path) = current_exe() {
                 Expression::String(path.to_str().unwrap().to_string())
             } else {
@@ -26,15 +26,15 @@ pub fn get() -> Expression {
         },
         String::from("license") => Expression::String("MIT".to_string()),
         String::from("prelude") => {
-            if let Some(home_dir) = dirs::home_dir() {
-                let prelude_path = home_dir.join("lumesh/config.lm");
+            if let Some(c) = dirs::config_dir() {
+                let prelude_path = c.join("lumesh/config.lm");
                 if prelude_path.exists() {
                     Expression::String(prelude_path.to_str().unwrap().to_string())
                 } else {
-                    Expression::None
+                    Expression::String(prelude_path.to_str().unwrap().to_string()+" !")
                 }
             } else {
-                Expression::None
+                Expression::String("config.lm".to_string())
             }
         }
     })